requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.12
langchain-openai==0.0.2.post1
pandas==2.1.4
pinecone-client==3.0.1
python-dotenv==1.0.0 